Finding collisions and decrypting cookie blocks is working!
Still have to clean up some things. We need to keep track of which
blocks we have already decrypted and then output the entire block
properly.

The "plain" property on the cookie block is repeated since each cookie
block exists in multipe "round trips" so we need to actually keep a
smaller array to track this.